split by comps
